// src/app/models/retiro_detail.ts
export interface RetiroDetail {
  art: number;
  can: number;
  can_ent?: number;
  can_sal?: number;
  com_fac: number;
  fch?: string;
  id?: number;
  plan_clt?: string;
  pre: number;
  prv: number;
  pre_pvp?: number;
  //emp: number;
  //emp_div: number;
  com_ped:number;
  com_ped_lin:number;
  usr_ord_plan?: string;
  usr_web?: string;
}